Stop using the Xml class in this codebase
This also fixes at least two escaping issues along the way. I found
messages that are added to the HTML without escaping. I think this
was a mistake. While technically possible it doesn't look like these
messages are supposed to contain raw HTML.

I try to replace open/close calls with a single rawElement call that
is more robust. It makes it easier to avoid mistakes like this.

Bug: T341775
Change-Id: Idc9e75205700f3212ef900a7285520a4689e4f5e